Create Custom_Toolbar_Seperators.as.css
A hack allowing for customizing and creating separators for the toolbar buttons. This is if you don't want to user the uc.js hack or if it eventually becomes obsolete. **NOTE** This WILL remove the "flexible space" option in the customization menu. You are essentially changing the flexible space into separators.